Add automated EnMasse installation playbooks
This adds two ansible playbooks. One for retrieving a released
version of the EnMasse playbook from GitHub and adding a playbook
into it that uses the 'local' host instead of using the 'enmasse'
host. This is currently run as a prerequisite of the other Ansible
playbook.

The second Ansible playbook will run the EnMasse playbook from the
location it was downloaded to.

The reason for two playbooks is due to the way calling one playbook
from another works in Ansible. It uses import_playbook. This task
is static though, meaning it's pre-processed and therefore checks
whether the playbook exists before running the play.

Since we need to download the EnMasse playbook before using
import_playbook it will always fail as the pre-processing can't
find the playbook on the machine.