Add user for admin work in Che/Launcher SSO
Che has an env var (CHE_SYSTEM_ADMIN_NAME) that can be set to give
a user with a matching username in Keycloak admin permissions in
Che. This user will be needed if we want to set permissions on
objects in Che like stacks (to allow all users to see and able to
use stacks they didn't create).

This change creates a user named admin and sets the env var to use
that name.