Remove LimitRange from che namespace
In some clusters there may be a default LimitRange created for each
project.
This change removes that limit for the Che namespace

This is for 2 reasons:

* The Che server liveness probe tends to timeout if the server doesn't
have enough resources.
* A Che java stack running a mvn command takes an extrememly long time
to run, and sometimes runs out of memory.

Removing this LimitRange prevents defaults from being applied to the
server & any workspaces (as they don't have specific resource requests
or limits set).

A future change may look at adding resource requests and limits on the
server & workspaces based on what's acutally needed.