INTLY-413 Lock launcher frontend and backend images to tags
Currently, for the fabric8/launcher-frontend and fabric8/launcher-
backend images we reference the master tag. This tag is updated
every time the master branch of their respective repos are updated.
This is an issue as it could lead to breaking changes being
introduced in an Integreatly release at a later date.

To resolve this, this change locks down the images to the latest
commit tags. These tags shouldn't be updated and removed, which
should make an Integreatly release more stable.

Verification:
* Run the installer
* Ensure the fabric8/launcher-backend image tag is b006e49
* Ensure the fabric8/launcher-frontend image tag is c885979