INTLY-778 Add uninstall playbook for the monitoring stack
Currently we only have an installer for the middleware monitoring
stack. However it would be good to be able to uninstall all of the
resources created by the installer in a simple way.

This change adds an uninstall playbook which removes every resource
created by the installer. It does this by re-creating the same
templates that are created by the installer, but invoking an `oc
delete` on them instead.

Verification:
- Run the install_monitoring_operator.yml playbook
- Run the uninstall_monitoring_operator.yml playbook
- Ensure the middleware-monitoring namespace is removed
- Ensure the other CRDS, ClusterRoles and ClusterRoleBindings
created by the installer are removed
- Try the same with a full uninstall (using uninstall.yml) to
ensure the overall uninstaller works too