INTLY-747 Include installer and uninstaller for managed fuse
Currently a user can provision their own instance of fuse. However
there isn't an already-running fuse instance available to them
when Integreatly starts up.

This adds a fuse which has it's subjectaccessreview configuration
removed so that any user who is authenticated in OpenShift can
access it.

This commit also includes the changes for the uninstallation of it

Verification (Preferably on a fresh cluster):
- Ensure the uninstaller is run if the cluster is not fresh
- Run the full installer
- Ensure a fuse namespace is created
- Login through the exposed route (syndesis-fuse...) as admin
- Login as a different user (an evals user)
- Ensure both logins worked successfully (you could try creating integrations
too or something if you want, to ensure fuse itself works)
- Run the uninstaller
- Ensure the fuse namespace is gone
- Ensure there are no fuse-ignite* imagestreams in the openshift namespace