Include node-exporter metrics in monitoring federation
We currently have federation between middleware-monitoring and
the Prometheus instance in openshift-monitoring. However, we only
are federating information about the Kube State Metrics metrics.

We should include node-exporter metrics in this too.

Verification:
- Run the installer
- Login to the middleware-monitoring Prometheus instance
- Ensure that the `openshift-monitoring-federation` target now
has two sets of match[] params on the targets page, one should
be `{ service="node-exporter" }`.
- Try to do a query like `{ service="node-exporter" }` and ensure
you are seeing a response that has data (a lot of entries), this
can take some time.