Move gitea pod count alert rule to it's own PrometheusRule in the gitea namespace
This change creates the GiteaPodCount alert only if Gitea is installed (true by default).
If the var is set to `gitea=false`, both Gitea and the GiteaPodCount alert will _not_ be installed.

The GiteaPodCount alert is moved from the `middleware_monitoring_config` role to it's own new role `gitea_config`