INTLY-1100 Add backup monitoring for CronJobs
Add a set of tasks for creating alerts that are triggered when
CronJobs and the Jobs they create are not in a correct state.

A summary of the alerts added in this commit:

- `JobRunningTimeExceeded` a single Job has been running for more
than five minutes, meaning it may be in debug mode in stuck.
- `CronJobSuspended` a CronJob that is being monitored has been
set to a suspended state, meaning Jobs won't be created.
- `CronJobNotRunInThreshold` a Job has not been created for a
CronJob in 25 hours, something may be wrong with creating the Job
- `CronJobExists_*` a CronJob does not exist in the cluster that
existed when the cluster was provisioned

All monitored CronJobs and Jobs are expected to have a label on
them named `monitoring-key` with the value `middleware`.

For a `CronJobExists_*` check to be created for a CronJob, the
CronJob must exist before the `monitoring.yml` tasks are run.

Install Verification:
- In `manifest.yml`, update `backup_resources_location` to point to
`https://raw.githubusercontent.com/aidenkeating/backup-container-image/INTLY-1100/templates/openshift`.
- In `manifest.yml`, update `backup_schedule` to `*/1 * * * *`.
- Run the managed install playbook.
- Ensure the install completes successfully.
- Ensure the created CronJobs have labels with the key
`monitoring-key` and the value `middleware`.
- Ensure that the Jobs created by the CronJobs have a label with
the key `monitoring-key` and the value `middleware`.
- Ensure the created CronJobs have the concurrency policy set to
`Forbid`.
- Go to the Prometheus route from the `middleware-monitoring`
namespace.
- Ensure their are the following alerts in the `Alerts` tab:
  - CronJobExists_* (One for each CronJob)
  - CronJobNotRunInThreshold
  - CronJobSuspended
  - JobRunningTimeExceeded

Alert Verification:
- Note that all alerts will take a few minutes to trigger as they
are based off of federated data which takes a while to sync
- Delete one of the CronJobs (3scale-mysql-backup) and ensure the
corresponding Alert is triggered
- `oc edit` one of the CronJobs (3scale-redis-backup), set
`.spec.suspend` to `true, save,` and ensure the CronJobSuspended
Alert is triggered for that CronJob
- `oc edit` one of the CronJobs (3scale-postgres-backup), set the
final line of `.spec.command` to `true` instead of being a blank
string. This should cause the next Job for that CronJob to go to
debug mode and sleep forever. Wait a few minutes and ensure the
`JobRunningTimeExceeded` Alert is triggered
- To test `CronJobNotRunInThreshold`, modify the PrometheusRule in
the `middleware-monitoring` namespace and change the `60*60*25`
(25 hours) to something small (`1`). You should see the Alert
triggering for the "sleeping" Job you created in the last step.
Note that due to federation, for small values, multiple Alerts
may trigger as the start times of jobs are out-of-date in our
Prometheus.