Use oc create in fuse managed instead of apply
Currently, we are using `oc apply` in many places in the Fuse
Managed installation. This causes issues when running the installation
over an existing installation as `oc apply` will update the existing
resources.

This change changes these places to `oc create` with a check that will
not fail if the resource already exists.

This change also includes an update to the Fuse Managed checks to see
if all Pods are up. Instead of referencing a label and specifying an
exact number of Pods, this will now specify a minimum amount of Pods
that should exist in the namespace before continuing.

Verification:
- Run the installer
- Ensure it completes successfully
  - Ensure the console can be accessed by any two users in the
    cluster, to verify the OAuth proxy rules are still correct
- Run the installer again, with the Fuse Managed namespace now
there
- Ensure it completes successfully, again