Prefix the webapp redirect URL in SSO if it's provided
Currently we don't take into account the ns_prefix when setting
the redirect URLs for the webapp in the cluster SSO. This results
in logging out of the webapp failing when the ns_prefix is set
during the installation.

This resolves this issue by taking the ns_prefix into account when
running the installer.