re-enable enmasse monitoring
This was set to false in this PR: https://github.com/integr8ly/installation/pull/1292. This was disabled during the AMQ 1.4 update as it broke the application monitoring stack due to the following upstream issue: https://github.com/EnMasseProject/enmasse/pull/4222

This has now been fixed upstream so we can re-enable the enmasse monitoring again.