I'm changing the flow of the upgrade in this commit.
The image stream upgrade should apply the same image stream templates as the install

One way would be to check what's in the install and copy it there.

Instead I am calling the main (install) from upgrade images and making that idempotent

The reason for that is so that the upgrade does not go stale over time if any of the templates change.

The version is pulled from the manifest file, if that's bumped it should cover install and upgrade.