As per the docs https://access.redhat.com/documentation/en-us/red_hat_amq/7.6/html/installing_and_managing_amq_online_on_openshift/upgrading-messaging
amq-online upgrade is completed by rerunning the installation.

we already have logic to download the artifact and install using the ansible playbook available

calling that logic here using import main in order to rerun the install.

The version for amq is already bumped previously for 1.7 so this should be pulled in as part of this upgrade(install)