ensure openshift_asset_url and openshift_master_url vars is available on osd
These vars are already being set on the osd group vars so we do not need to set them here. This should only be set if 'run_master_tasks' var is set to true which is the case for pds. This var is set to false by default on osd