Removing the heimdall_pull_secret_name environment variable
The username (which this variable was being used for in the template)
is  embedded in the pull secret.

When the secret is created in openshift the username and token is present

The pull secret name is hardcoded. It should stay this way as the
deployment expects a secret with the name heimdall-dockercfg

https://github.com/integr8ly/heimdall/blob/master/deploy/operator.yaml#L50