Add custom SOP URLs to enmasse alert mails
JIRA: https://issues.redhat.com/browse/INTLY-8385

This changes the email template used by AlertManager to check if
the alert being mailed is one of a particular set (which happen to not
have sop_urls defined), and if it is, adds a sop_url.

Additionally, there are two alerts that have a severity of "warning",
but CS-SRE expect to be notified about them as critical, so this
change also adds a specific alert matcher to the route for the
critical receiver.

The alternative to these changes would be to suppress the existing
alerts, and duplicate them into new alerts that have the correct
severity and sop_url annotations set. This way is considered the
lesser evil. :)